feat: Add FederatedTTestIndependent

+++ exaflow/algorithms/exareme3/library/ttest_common.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel
 
 class TTestResult(BaseModel):
     t_stat: float
-    df: int
+    df: float
     p: float
     mean_diff: float
     se_diff: float
